rename NewsFormStyles file to FormStyle
